+++ src/Service/Metadata/Manager.php
@@ -38,6 +38,10 @@ final class Manager implements ManagerInterface
     {
         $metadata = $this->fileExportedParser->parse($this->metadataPath);
 
+        if (empty($metadata)) {
+            throw new EmptyMetadataException('Should not apply empty metadata.');
+        }
+
         $this->apiClient->metadata()->replace($metadata, $allowInconsistency);
     }