+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 + Bug #13: Fix cast object as array when apply metadata (insert, update, delete table permissions check and filter fields).
 + Bug #15: Re-export metadata had been removed not affect.
++ Enh #16: Should throw exception when apply empty metadata.
 
 ## 1.0.2